requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
aiofiles>=23.2.0
motor>=3.4.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# LLM providers — install the one(s) you need
groq>=0.9.0
google-generativeai>=0.7.0
anthropic>=0.25.0
